update.
M  .config/BraveSoftware/Brave-Browser/Default/Bookmarks
M  .config/LSD/.harpoon_on_steroids_data.txt
M  .config/LSD/exit.sh
M  .config/LSD/harpoon_on_steroids.sh
M  .config/i3/config
M  .config/mimeapps.list
M  .config/picom/picom.conf
M  .config/polybar/config.ini
M  .config/polybar/launch.sh
M  .zshrc

+++ .config/polybar/launch.sh
@@ -1,14 +1,12 @@
 #!/usr/bin/env bash
 
 # Add this script to your wm startup file.
-
 DIR="$HOME/.config/polybar"
 
-# Terminate already running bar instances
-killall -q polybar
-
-# Wait until the processes have been shut down
-while pgrep -u $UID -x polybar >/dev/null; do sleep 1; done
-
-# Launch the bar
-polybar mybar
+if pgrep -x "polybar" > /dev/null; then
+    killall -q polybar
+else
+    # Wait until the processes have been shut down
+    while pgrep -u $UID -x polybar >/dev/null; do sleep 1; done
+    polybar mybar &
+fi
